% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DNAm_Age.R
\name{do_dnam_clock_mouse}
\alias{do_dnam_clock_mouse}
\title{Calculation function for mouse DNAm age}
\usage{
do_dnam_clock_mouse(
  data,
  clock = c("MeerClock", "PetkovichClock", "ThompsonClock", "WangLiver"),
  fuzzy_pos_window = 100
)
}
\arguments{
\item{data}{A dataframe with the chr_pos id as first column. The rest of the columns should be methylation levels for each sample, with sample names as column names.}

\item{clock}{A string vector with the names of the clocks to use.}

\item{fuzzy_pos_window}{A interger indicating a window size that to be used to infer missing methylation value. Default is 100.}
}
\value{
An tibble with sample per row and DNAm age as columns.
}
\description{
Calculation function for mouse DNAm age
}
\examples{
do_dnam_clock_mouse(df, c("MeerClock", "PetkovichClock", "ThompsonClock", "WangLiver"), 100)
}
